added datetime test/fix to sample
