+++ tests/unit/serialization_utils_test.py
@@ -1,8 +1,10 @@
 import pytest
 import pandas as pd
 from buckaroo.serialization_utils import df_to_obj
-from pandas.core.dtypes.dtypes import DatetimeTZDtype
-
+from buckaroo.customizations.down_sample import (
+    is_col_dt_safe, is_dataframe_datetime_safe, check_and_fix_df,
+    DuplicateColumnsException
+)
 
 def test_df_to_obj():
     named_index_df = pd.DataFrame(
@@ -12,46 +14,6 @@ def test_df_to_obj():
     serialized_df = df_to_obj(named_index_df, {})
     assert serialized_df['data'][0]['names'] == 'one'
 
-from pandas._libs.tslibs import timezones
-def is_col_dt_safe(col_or_index):
-    if isinstance(col_or_index.dtype, DatetimeTZDtype):
-        dt = col_or_index.dtype
-        if timezones.is_utc(dt.tz):
-            return True
-        elif hasattr(dt.tz, 'zone'):
-            return True
-        return False
-    return True
-
-def is_dataframe_datetime_safe(df):
-    for col in df:
-        if not is_col_dt_safe(df[col]):
-            return False
-    if not is_col_dt_safe(df.index):
-        return False
-    return True
-
-def fix_df_dates(df):
-    for col in df:
-        if not is_col_dt_safe(df[col]):
-            print("col", col)
-            df[col] = df[col].tz_convert('UTC')
-    if not is_col_dt_safe(df.index):
-        df.index = df.index.tz_convert('UTC')
-    return df
-
-class DuplicateColumnsException(Exception):
-    pass
-
-
-def check_and_fix_df(df):
-    if not df.columns.is_unique:
-        raise DuplicateColumnsException("Your dataframe has duplicate columns. Buckaroo requires distinct column names")
-    if not is_dataframe_datetime_safe(df):
-        print("your dataframe has a column or index with a datetime series without atimezone.  Setting a default UTC timezone to proceed with display. https://github.com/paddymul/buckaroo/issues/277")
-        return fix_df_dates(df)
-    df.to_json(orient='table')
-    return df
 
 dt_strs = ['2024-06-24 09:32:00-04:00', '2024-06-24 09:33:00-04:00', '2024-06-24 09:34:00-04:00']
 dt_series = pd.Series(pd.to_datetime(dt_strs).values)
